requirements.txt
pandas
numpy
matplotlib
seaborn
tqdm
sentence-transformers
faiss-cpu
scikit-learn
gradio
transformers
pickle5 